+++ src/main/java/com/alper/finance/entity/NetWorthType.java
@@ -1,0 +1,8 @@
+package com.alper.finance.entity;
+
+public enum NetWorthType {
+    ASSET,
+    LIABILITY
+}
+
+